--- requirement-generator/checkers/colleges/bu.ts
@@ -1,1 +1,0 @@
-export default {};
