--- requirement-generator/checkers/colleges/he.ts
@@ -1,5 +1,0 @@
-import { Course } from '../../types';
-
-const humanEcologyCredits = (course: Course): boolean => course.acadGroup.includes('HE');
-
-export default { humanEcologyCredits };
