--- requirement-generator/checkers/types.ts
@@ -1,4 +1,0 @@
-import { Course } from '../types';
-
-export type Checker = (course: Course) => boolean;
-export type Checkers = { readonly [checkerName: string]: Checker | readonly Checker[] };
