+++ src/requirements/checkers/colleges/bu.ts
@@ -1,0 +1,1 @@
+export default {};
