+++ src/requirements/checkers/colleges/he.ts
@@ -1,0 +1,5 @@
+import { Course } from '../../types';
+
+const humanEcologyCredits = (course: Course): boolean => course.acadGroup.includes('HE');
+
+export default { humanEcologyCredits };
