+++ src/requirements/checkers/types.ts
@@ -1,0 +1,4 @@
+import { Course } from '../types';
+
+export type Checker = (course: Course) => boolean;
+export type Checkers = { readonly [checkerName: string]: Checker | readonly Checker[] };
