+++ src/requirements/types.ts
@@ -3,13 +3,11 @@ export type Course = {
   readonly catalogNbr: string;
   readonly titleLong: string;
   readonly description: string;
-  readonly catalogBreadth: string;
-  readonly catalogDistr: string;
-  readonly catalogLang: string;
+  readonly catalogBreadth?: string;
+  readonly catalogDistr?: string;
   readonly catalogAttribute: string;
-  readonly catalogWhenOffered: string;
-  readonly catalogComments: string;
-  readonly catalogSatisfiesReq: string;
+  readonly catalogComments?: string;
+  readonly catalogSatisfiesReq?: string;
   readonly acadCareer: string;
   readonly acadGroup: string;
 };
@@ -21,8 +19,6 @@ export type CourseTaken = {
   readonly number: string;
   readonly credits: number;
 };
-
-export type RequirementSearch = keyof Course | 'all' | 'self-check' | 'all-eligible' | 'code';
 
 export interface BaseRequirement {
   readonly name: string;
@@ -40,6 +36,10 @@ export type UniversityRequirements = {
   readonly requirements: readonly BaseRequirement[];
 };
 
+export interface CollegeOrMajorRequirement extends BaseRequirement {
+  readonly checkerName: string | null;
+}
+
 export type EligibleCourses = {
   readonly [semester: string]: {
     // Subjects to course numbers
@@ -47,27 +47,35 @@ export type EligibleCourses = {
   };
 };
 
-export interface CollegeOrMajorRequirement extends BaseRequirement {
+export interface DecoratedCollegeOrMajorRequirement extends BaseRequirement {
   readonly courses: readonly EligibleCourses[];
 }
 
-export type CollegeRequirements = {
+export type CollegeRequirements<R> = {
   readonly [collegeCode: string]: {
     readonly name: string;
-    readonly requirements: readonly CollegeOrMajorRequirement[];
+    readonly requirements: readonly R[];
   };
 };
 
-export type MajorRequirements = {
+export type MajorRequirements<R> = {
   readonly [collegeCode: string]: {
     readonly name: string;
     readonly schools: readonly string[];
-    readonly requirements: readonly CollegeOrMajorRequirement[];
+    readonly requirements: readonly R[];
   };
 };
 
-export type RequirementsJson = {
+type GenericRequirementsJson<R> = {
   readonly university: UniversityRequirements;
-  readonly college: CollegeRequirements;
-  readonly major: MajorRequirements;
+  readonly college: CollegeRequirements<R>;
+  readonly major: MajorRequirements<R>;
+};
+
+export type RequirementsJson = GenericRequirementsJson<CollegeOrMajorRequirement>;
+
+export type DecoratedRequirementsJson = {
+  readonly university: UniversityRequirements;
+  readonly college: CollegeRequirements<DecoratedCollegeOrMajorRequirement>;
+  readonly major: MajorRequirements<DecoratedCollegeOrMajorRequirement>;
 };
